showcase_local_coffee_shop/google_cloud_run/load_to_bq/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery
